First experiments with PseudoRansac. Found some adequate parameters
Tracker is slow and does not track as necessary. Added more complicated
test for MeanShift.

+++ src/MeanShift3D.hpp
@@ -8,7 +8,7 @@ class MeanShift3D
 {
 public:
     MeanShift3D(
-        const size_t MaxIter,
+        const int MaxIter,
         const double eps,
         const cv::Point3f windowSizes,
         const bool isIterCriteria = true);
@@ -18,7 +18,7 @@ public:
         cv::Mat& foundCenter) const;
 
 private:
-    size_t maxIter;
+    int maxIter;
     double  eps;
     cv::Point3f windowSizes;
     bool isIterCriteria;
